Move DivvunSpell supporting classes into files
